Turning off checks on indentation
Currently scalariform is unable to format code that will be compliant
with scalastyle indentation checks.

The issue is that scalastyle checks that multiline constructor arguments
are indented with 4 spaces instead of 2 (as written in the scala style
guideline), but scalariform doesn’t format that correctly.

These are 2 pending pull requests that should solve this issue, but the
last activity on them seem to be in September of last year:
- scala-ide/scalariform#227
- scala-ide/scalariform#192

Another option would be to fine tune scalastyle to not complain about
this specific issue but even though there are some configuration
available in terms of indentation (for example for multiline method
parameter) there seems to be no configuration for multiline constructor
parameter. So my suggestion is that we turn off the indentation checks.
If we compile and test the code always using maven the code will be
always automatically formatted by scalariform anyway (because it runs in
a previous phase in maven) so the indentation will be okay.